catch up with the main site.

+++ tests/travis/run_test.sh
@@ -5,28 +5,25 @@ make -f dmlc-core/scripts/packages.mk lz4
 
 if [ ${TRAVIS_OS_NAME} == "osx" ]; then
     echo 'USE_OPENMP=0' >> config.mk
-    echo 'TMPVAR := $(XGB_PLUGINS)' >> config.mk
-    echo 'XGB_PLUGINS = $(filter-out plugin/lz4/plugin.mk, $(TMPVAR))' >> config.mk
-else
-    # use g++-4.8 for linux
-    export CXX=g++-4.8
 fi
 
 if [ ${TASK} == "python_test" ]; then
     make all || exit -1
     echo "-------------------------------"
-    source activate python3
+    conda activate python3
     python --version
     conda install numpy scipy pandas matplotlib scikit-learn
 
     python -m pip install graphviz pytest pytest-cov codecov
-    python -m pip install https://h2o-release.s3.amazonaws.com/datatable/stable/datatable-0.7.0/datatable-0.7.0-cp37-cp37m-linux_x86_64.whl
+    python -m pip install dask distributed dask[dataframe]
+    python -m pip install datatable
     python -m pytest -v --fulltrace -s tests/python --cov=python-package/xgboost || exit -1
     codecov
 fi
 
 if [ ${TASK} == "java_test" ]; then
-    set -e
+    export RABIT_MOCK=ON
+    conda activate python3
     cd jvm-packages
     mvn -q clean install -DskipTests -Dmaven.test.skip
     mvn -q test
@@ -34,20 +31,18 @@ fi
 
 if [ ${TASK} == "cmake_test" ]; then
     set -e
-    # Build gtest via cmake
-    wget -nc https://github.com/google/googletest/archive/release-1.7.0.zip
-    unzip -n release-1.7.0.zip
-    mv googletest-release-1.7.0 gtest && cd gtest
-    CC=gcc-7 CXX=g++-7 cmake . && make
-    mkdir lib && mv libgtest.a lib
-    cd ..
-    rm -rf release-1.7.0.zip
+
+    if grep -n -R '<<<.*>>>\(.*\)' src include | grep --invert "NOLINT"; then
+        echo 'Do not use raw CUDA execution configuration syntax with <<<blocks, threads>>>.' \
+             'try `dh::LaunchKernel`'
+        exit -1
+    fi
 
     # Build/test
     rm -rf build
     mkdir build && cd build
     PLUGINS="-DPLUGIN_LZ4=ON -DPLUGIN_DENSE_PARSER=ON"
-    CC=gcc-7 CXX=g++-7 cmake .. -DGOOGLE_TEST=ON -DGTEST_ROOT=$PWD/../gtest/ ${PLUGINS}
+    CC=gcc-9 CXX=g++-9 cmake .. -DGOOGLE_TEST=ON -DUSE_OPENMP=ON -DUSE_DMLC_GTEST=ON ${PLUGINS}
     make
     ./testxgboost
     cd ..
